Refactor field filtering into seperate filtering package
