+++ src/fhir_mcp_server/utils.py
@@ -16,7 +16,6 @@
 
 import aiohttp
 import logging
-import fhirpathpy
 
 from fhir_mcp_server.oauth import ServerConfigs
 
@@ -125,144 +124,6 @@ async def get_capability_statement(metadata_url: str) -> Dict[str, Any]:
         )
         raise ValueError("Unable to fetch FHIR metadata")
 
-
-def _split_bundle_resource_paths(
-    field_paths: List[str],
-) -> tuple[List[str], List[str]]:
-    """Split field_paths into two buckets:
-
-    - bundle_paths: Bundle.* paths applied to the Bundle wrapper only
-    - resource_paths: paths applied to each entry resource — includes Bundle.entry.resource.* (prefix stripped) and resource-type paths (e.g. Patient.name)
-    """
-    bundle_paths = []
-    resource_paths = []
-    for path in field_paths:
-        if path.startswith("Bundle.entry.resource."):
-            resource_paths.append(path.removeprefix("Bundle.entry.resource."))
-        elif path.startswith("Bundle."):
-            bundle_paths.append(path)
-        else:
-            resource_paths.append(path)
-    return bundle_paths, resource_paths
-
-
-def _filter_with_fhirpath(
-    resource: Dict[str, Any],
-    field_paths: List[str],
-) -> Dict[str, Any]:
-    """Apply FHIRPath expressions to a single FHIR resource."""
-
-    resource_type = resource.get("resourceType", "")
-
-    result: Dict[str, Any] = {}
-    not_matched: List[str] = []
-    errors: List[str] = []
-
-    for expr in field_paths:
-        prefix = expr.split(".")[0]
-        if not prefix or not prefix[0].isupper():
-            not_matched.append(expr)
-            continue
-        if prefix != resource_type:
-            continue
-        try:
-            matched = fhirpathpy.evaluate(resource, expr)
-            if matched:
-                key = expr.removeprefix(f"{resource_type}.")
-                top_level_field = key.split(".")[0].split("(")[0]
-                is_array = isinstance(resource.get(top_level_field), list)
-                result[key] = matched if len(matched) > 1 or is_array else matched[0]
-            else:
-                not_matched.append(expr)
-        except Exception as e:
-            logger.warning("FHIRPath eval failed for expression %r: %s", expr, e)
-            errors.append(expr)
-
-    logger.debug(
-        "%s: matched %s, skipped %s",
-        resource_type or "unknown",
-        list(result),
-        not_matched + errors,
-    )
-    if not_matched:
-        result["_not_matched"] = not_matched
-    if errors:
-        result["_errors"] = errors
-    return result
-
-
-def _filter_entry(
-    entry: Dict[str, Any],
-    paths: List[str],
-    is_search: bool,
-) -> Dict[str, Any]:
-    if "resource" not in entry or not isinstance(entry["resource"], dict):
-        return _filter_with_fhirpath(entry, paths)
-
-    resource = entry["resource"]
-    search_mode = entry.get("search", {}).get("mode")
-    resource_type = resource.get("resourceType", "")
-
-    # e.g. searchset Bundle → entry → collection Bundle → filter its entries too
-    if resource_type == "Bundle":
-        _, inner_paths = _split_bundle_resource_paths(paths)
-        inner_entries = [
-            _filter_with_fhirpath(inner_entry["resource"], inner_paths)
-            if "resource" in inner_entry and isinstance(inner_entry["resource"], dict)
-            else inner_entry
-            for inner_entry in resource.get("entry", [])
-        ]
-        result = {"entry": inner_entries}
-        if search_mode == "include":
-            result["search"] = entry["search"]
-        return result
-
-    if not is_search or search_mode == "include":
-        always_include = [f"{resource_type}.id", f"{resource_type}.resourceType"]
-    elif search_mode == "match":
-        always_include = [f"{resource_type}.id"]
-    else:
-        always_include = []
-
-    result = _filter_with_fhirpath(resource, always_include + paths)
-    if search_mode == "include":
-        result["search"] = entry["search"]
-    return result
-
-
-def _process_one_resource(
-    item: Any,
-    field_paths: List[str],
-    is_search: bool,
-) -> Any:
-    if not isinstance(item, dict):
-        return item
-
-    if item.get("resourceType") == "Bundle":
-        bundle_paths, resource_paths = _split_bundle_resource_paths(field_paths)
-        result = _filter_with_fhirpath(item, bundle_paths) if bundle_paths else {}
-        if resource_paths:
-            result["entry"] = [
-                _filter_entry(entry, resource_paths, is_search)
-                for entry in item.get("entry", [])
-            ]
-        return result
-
-    return _filter_entry(item, field_paths, is_search)
-
-
-def filter_resource_fields(
-    data: Any,
-    field_paths: List[str] | None = None,
-    is_search: bool = False,
-) -> Any:
-    if not field_paths:
-        return data
-
-    if isinstance(data, list):
-        return [_process_one_resource(item, field_paths, is_search) for item in data]
-
-    return _process_one_resource(data, field_paths, is_search)
 
 
 def get_default_headers() -> Dict[str, str]:
